refactor: page structure; maintain reserved usernames
